Fix bug when changing between questions with unequal number of parts
Changing between two questions would cause an effect to run with the new
parts and the old answers. If the old question had more parts than the new
question, this would result in a indexing error. Fixed this by allowing the
effect where this was happening to only run after the answers are initialized
for the new question.